server/requirements.txt
fastapi
finnhub_python
python-dotenv
yfinance
psycopg2-binary